Move JSONAPISerializer to jsonapi/serializer.js
