+++ build-support/globalize-orbit-common-jsonapi.js
@@ -1,2 +1,2 @@
 window.OC.JSONAPISource = requireModule("orbit-common/jsonapi-source")["default"];
-window.OC.JSONAPISerializer = requireModule("orbit-common/jsonapi-serializer")["default"];
+window.OC.JSONAPISerializer = requireModule("orbit-common/jsonapi/serializer")["default"];
