+++ lib/orbit-common/jsonapi-source.js
@@ -10,7 +10,7 @@ import Updatable from 'orbit/updatable';
 import TransformBuilder from './transform/builder';
 import Source from './source';
 import Serializer from './serializer';
-import JSONAPISerializer from './jsonapi-serializer';
+import JSONAPISerializer from './jsonapi/serializer';
 import { getUpdateRequests, UpdateRequestProcessors } from './jsonapi/update-requests';
 import { OperationNotAllowed, TransformNotAllowed, RecordNotFoundException, RecordAlreadyExistsException } from './lib/exceptions';
 import CacheIntegrityProcessor from 'orbit-common/cache/operation-processors/cache-integrity-processor';
