+++ lib/orbit-common/jsonapi/serializer.js
@@ -1,7 +1,7 @@
-import Serializer from './serializer';
+import Serializer from '../serializer';
 import { clone, isArray, isObject } from 'orbit/lib/objects';
 import { dasherize, camelize } from 'orbit/lib/strings';
-import { parseIdentifier, toIdentifier } from 'orbit-common/lib/identifiers';
+import { parseIdentifier, toIdentifier } from '../lib/identifiers';
 
 export default class JSONAPISerializer extends Serializer {
   resourceKey(type) {
