+++ test/tests/orbit-common/unit/jsonapi-serializer-test.js
@@ -3,7 +3,7 @@ import Orbit from 'orbit/main';
 import { uuid } from 'orbit/lib/uuid';
 import Schema from 'orbit-common/schema';
 import Serializer from 'orbit-common/serializer';
-import JSONAPISerializer from 'orbit-common/jsonapi-serializer';
+import JSONAPISerializer from 'orbit-common/jsonapi/serializer';
 import { parseIdentifier, toIdentifier } from 'orbit-common/lib/identifiers';
 import { Promise } from 'rsvp';
